small datetime bug fix on form edit
